# [Tech Request]: fix stats for values scan