@jensenojs 2024-01-10
手动在本地跑了tpcc和sysbench的场景, sysbench和main分支的性能差距并不明显, tpcc甚至还要好一些, 请QA测试一下, 如果没有明显问题的话就重新打开统计好了.

# tpcc 

## this pr

2024-01-10 10:58:12 INFO  jTPCC:803 - Term-00, Term-00, Running Average tpmTOTAL: 6359.01    Current tpmTOTAL: 210456    Memory Usage: 59MB / 256MB          
2024-01-10 10:58:12 INFO  jTPCC:793 - Term-00, All terminals finished executing 2024-01-10 10:58:12

2024-01-10 10:58:12 INFO  jTPCC:761 - Term-00, 
2024-01-10 10:58:12 INFO  jTPCC:762 - Term-00, 
2024-01-10 10:58:12 INFO  jTPCC:763 - Term-00, Measured tpmC (NewOrders) = 2859.0
2024-01-10 10:58:12 INFO  jTPCC:764 - Term-00, Measured tpmTOTAL = 6356.77
2024-01-10 10:58:12 INFO  jTPCC:765 - Term-00, Measured tpmE (ErrorCount) = 0.0
2024-01-10 10:58:12 INFO  jTPCC:766 - Term-00, Session Start     = 2024-01-10 10:53:12
2024-01-10 10:58:12 INFO  jTPCC:767 - Term-00, Session End       = 2024-01-10 10:58:12
2024-01-10 10:58:12 INFO  jTPCC:768 - Term-00, Transaction Count = 31805
2024-01-10 10:58:12 INFO  jTPCC:769 - Term-00, Transaction Error = 0
2024-01-10 10:58:12 INFO  jTPCC:770 - Term-00, Transaction NewOrders = 14305
2024-01-10 10:58:12 INFO  jTPCC:793 - Term-00, Session finished!


## main (base on 4541bf1364e)

2024-01-10 11:05:33 INFO  jTPCC:803 - Term-00, Term-00, Running Average tpmTOTAL: 5104.31    Current tpmTOTAL: 169632    Memory Usage: 50MB / 256MB          
2024-01-10 11:05:33 INFO  jTPCC:793 - Term-00, All terminals finished executing 2024-01-10 11:05:33

2024-01-10 11:05:33 INFO  jTPCC:761 - Term-00, 
2024-01-10 11:05:33 INFO  jTPCC:762 - Term-00, 
2024-01-10 11:05:33 INFO  jTPCC:763 - Term-00, Measured tpmC (NewOrders) = 2284.01
2024-01-10 11:05:33 INFO  jTPCC:764 - Term-00, Measured tpmTOTAL = 5104.2
2024-01-10 11:05:33 INFO  jTPCC:765 - Term-00, Measured tpmE (ErrorCount) = 0.0
2024-01-10 11:05:33 INFO  jTPCC:766 - Term-00, Session Start     = 2024-01-10 11:00:33
2024-01-10 11:05:33 INFO  jTPCC:767 - Term-00, Session End       = 2024-01-10 11:05:33
2024-01-10 11:05:33 INFO  jTPCC:768 - Term-00, Transaction Count = 25531
2024-01-10 11:05:33 INFO  jTPCC:769 - Term-00, Transaction Error = 0
2024-01-10 11:05:33 INFO  jTPCC:770 - Term-00, Transaction NewOrders = 11425


# sysbench


## this pr

### read only 

SQL statistics:
    queries performed:
        read:                            892861
        write:                           0
        other:                           0
        total:                           892861
    transactions:                        892861 (29760.88 per sec.)
    queries:                             892861 (29760.88 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          30.0010s
    total number of events:              892861

Latency (ms):
         min:                                    0.09
         avg:                                    0.34
         max:                                  107.57
         95th percentile:                        0.00
         sum:                               299885.78

Threads fairness:
    events (avg/stddev):           89286.1000/516.52
    execution time (avg/stddev):   29.9886/0.00


### point_select

SQL statistics:
    queries performed:
        read:                            927393
        write:                           0
        other:                           0
        total:                           927393
    transactions:                        927393 (30911.79 per sec.)
    queries:                             927393 (30911.79 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          30.0010s
    total number of events:              927393

Latency (ms):
         min:                                    0.09
         avg:                                    0.32
         max:                                   85.94
         95th percentile:                        0.00
         sum:                               299871.01

Threads fairness:
    events (avg/stddev):           92739.3000/388.25
    execution time (avg/stddev):   29.9871/0.00

### insert
SQL statistics:
    queries performed:
        read:                            0
        write:                           49800
        other:                           0
        total:                           49800
    transactions:                        49800  (1659.77 per sec.)
    queries:                             49800  (1659.77 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          30.0038s
    total number of events:              49800

Latency (ms):
         min:                                    1.66
         avg:                                    6.02
         max:                                   68.47
         95th percentile:                        0.00
         sum:                               299890.44

Threads fairness:
    events (avg/stddev):           4980.0000/6.43
    execution time (avg/stddev):   29.9890/0.00


### update_index

SQL statistics:
    queries performed:
        read:                            0
        write:                           0
        other:                           615519
        total:                           615519
    transactions:                        615519 (20516.55 per sec.)
    queries:                             615519 (20516.55 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          30.0008s
    total number of events:              615519

Latency (ms):
         min:                                    0.23
         avg:                                    0.49
         max:                                   36.89
         95th percentile:                        0.00
         sum:                               299881.82

Threads fairness:
    events (avg/stddev):           61551.9000/123.04
    execution time (avg/stddev):   29.9882/0.00

## main (base on 4541bf1364e)

### read only

SQL statistics:
    queries performed:
        read:                            911918
        write:                           0
        other:                           0
        total:                           911918
    transactions:                        911918 (30395.89 per sec.)
    queries:                             911918 (30395.89 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          30.0011s
    total number of events:              911918

Latency (ms):
         min:                                    0.11
         avg:                                    0.33
         max:                                   61.73
         95th percentile:                        0.00
         sum:                               299858.51

Threads fairness:
    events (avg/stddev):           91191.8000/382.36
    execution time (avg/stddev):   29.9859/0.00

### point_select

SQL statistics:
    queries performed:
        read:                            919084
        write:                           0
        other:                           0
        total:                           919084
    transactions:                        919084 (30634.86 per sec.)
    queries:                             919084 (30634.86 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          30.0010s
    total number of events:              919084

Latency (ms):
         min:                                    0.11
         avg:                                    0.33
         max:                                   56.17
         95th percentile:                        0.00
         sum:                               299851.56

Threads fairness:
    events (avg/stddev):           91908.4000/338.75
    execution time (avg/stddev):   29.9852/0.00

### insert

SQL statistics:
    queries performed:
        read:                            0
        write:                           50036
        other:                           0
        total:                           50036
    transactions:                        50036  (1667.59 per sec.)
    queries:                             50036  (1667.59 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          30.0044s
    total number of events:              50036

Latency (ms):
         min:                                    1.40
         avg:                                    5.99
         max:                                  113.94
         95th percentile:                        0.00
         sum:                               299884.00

Threads fairness:
    events (avg/stddev):           5003.6000/6.76
    execution time (avg/stddev):   29.9884/0.00

### update_index

SQL statistics:
    queries performed:
        read:                            0
        write:                           0
        other:                           595944
        total:                           595944
    transactions:                        595944 (19863.88 per sec.)
    queries:                             595944 (19863.88 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

General statistics:
    total time:                          30.0012s
    total number of events:              595944

Latency (ms):
         min:                                    0.23
         avg:                                    0.50
         max:                                   52.17
         95th percentile:                        0.00
         sum:                               299877.97

Threads fairness:
    events (avg/stddev):           59594.4000/82.05
    execution time (avg/stddev):   29.9878/0.00